+++ include/openrave/interface.h
@@ -91,6 +91,9 @@ public:
     /// \brief Returns the readable interface. <b>[multi-thread safe]</b>
     virtual ReadablePtr GetReadableInterface(const std::string& id) const;
 
+    /// \brief Returns whether a readable interface exists. <b>[multi-thread safe]</b>
+    virtual bool HasReadableInterface(const std::string& id) const;
+
     /// \brief Set a new readable interface and return the previously set interface if it exists. <b>[multi-thread safe]</b>
     virtual ReadablePtr SetReadableInterface(const std::string& id, ReadablePtr readable);